test(test-utils.tsx): renamed test-utils.ts to .tsx and added router context
Still not used in code
